+++ SeaBot/Data/Defenitions/BuildingDefenitions.cs
@@ -34,18 +34,18 @@ namespace SeaBotCore.Data.Defenitions
 
         public class Item
         {
-            [J("def_id")] public long DefId { get; set; }
-            [J("version_id")] public long VersionId { get; set; }
+            [J("def_id")] public int DefId { get; set; }
+            [J("version_id")] public int VersionId { get; set; }
             [J("name")] public string Name { get; set; }
             [J("name_loc")] public string NameLoc { get; set; }
             [J("desc_loc")] public string DescLoc { get; set; }
-            [J("tile_width")] public long TileWidth { get; set; }
-            [J("tile_height")] public long TileHeight { get; set; }
-            [J("on_grid")] public long OnGrid { get; set; }
+            [J("tile_width")] public int TileWidth { get; set; }
+            [J("tile_height")] public int TileHeight { get; set; }
+            [J("on_grid")] public int OnGrid { get; set; }
             [J("type")] public string Type { get; set; }
-            [J("max_level")] public long MaxLevel { get; set; }
-            [J("hide")] public long Hide { get; set; }
-            [J("upgrade_priority")] public long UpgradePriority { get; set; }
+            [J("max_level")] public int MaxLevel { get; set; }
+            [J("hide")] public int Hide { get; set; }
+            [J("upgrade_priority")] public int UpgradePriority { get; set; }
             [J("levels")] public Levels Levels { get; set; }
         }
 
@@ -56,38 +56,38 @@ namespace SeaBotCore.Data.Defenitions
 
         public class Level
         {
-            [J("id")] public long Id { get; set; }
+            [J("id")] public int Id { get; set; }
             [J("req_type")] public string ReqType { get; set; }
-            [J("req_id")] public long ReqId { get; set; }
-            [J("req_level")] public long ReqLevel { get; set; }
-            [J("player_level")] public long PlayerLevel { get; set; }
-            [J("player_level_gem")] public long PlayerLevelGem { get; set; }
-            [J("xp")] public long Xp { get; set; }
-            [J("sailors_regen")] public long SailorsRegen { get; set; }
-            [J("sailors_cap")] public long SailorsCap { get; set; }
-            [J("gem_price")] public long GemPrice { get; set; }
-            [J("gem_upgrade_time")] public long GemUpgradeTime { get; set; }
-            [J("upgrade_time")] public long UpgradeTime { get; set; }
-            [J("boat_level")] public long BoatLevel { get; set; }
-            [J("lighthouse_level")] public long LighthouseLevel { get; set; }
-            [J("museum_level")] public long MuseumLevel { get; set; }
-            [J("iso_x")] public long IsoX { get; set; }
-            [J("iso_y")] public long IsoY { get; set; }
+            [J("req_id")] public int ReqId { get; set; }
+            [J("req_level")] public int ReqLevel { get; set; }
+            [J("player_level")] public int PlayerLevel { get; set; }
+            [J("player_level_gem")] public int PlayerLevelGem { get; set; }
+            [J("xp")] public int Xp { get; set; }
+            [J("sailors_regen")] public int SailorsRegen { get; set; }
+            [J("sailors_cap")] public int SailorsCap { get; set; }
+            [J("gem_price")] public int GemPrice { get; set; }
+            [J("gem_upgrade_time")] public int GemUpgradeTime { get; set; }
+            [J("upgrade_time")] public int UpgradeTime { get; set; }
+            [J("boat_level")] public int BoatLevel { get; set; }
+            [J("lighthouse_level")] public int LighthouseLevel { get; set; }
+            [J("museum_level")] public int MuseumLevel { get; set; }
+            [J("iso_x")] public int IsoX { get; set; }
+            [J("iso_y")] public int IsoY { get; set; }
             [J("bounds")] public string Bounds { get; set; }
-            [J("iso_width")] public long IsoWidth { get; set; }
-            [J("iso_height")] public long IsoHeight { get; set; }
-            [J("pivot_x")] public long PivotX { get; set; }
-            [J("pivot_y")] public long PivotY { get; set; }
+            [J("iso_width")] public int IsoWidth { get; set; }
+            [J("iso_height")] public int IsoHeight { get; set; }
+            [J("pivot_x")] public int PivotX { get; set; }
+            [J("pivot_y")] public int PivotY { get; set; }
             [J("texture")] public string Texture { get; set; }
             [J("particles")] public string Particles { get; set; }
             [J("animations")] public string Animations { get; set; }
-            [J("iso_x_new")] public long IsoXNew { get; set; }
-            [J("iso_y_new")] public long IsoYNew { get; set; }
+            [J("iso_x_new")] public int IsoXNew { get; set; }
+            [J("iso_y_new")] public int IsoYNew { get; set; }
             [J("bounds_new")] public string BoundsNew { get; set; }
-            [J("iso_width_new")] public long IsoWidthNew { get; set; }
-            [J("iso_height_new")] public long IsoHeightNew { get; set; }
-            [J("pivot_x_new")] public long PivotXNew { get; set; }
-            [J("pivot_y_new")] public long PivotYNew { get; set; }
+            [J("iso_width_new")] public int IsoWidthNew { get; set; }
+            [J("iso_height_new")] public int IsoHeightNew { get; set; }
+            [J("pivot_x_new")] public int PivotXNew { get; set; }
+            [J("pivot_y_new")] public int PivotYNew { get; set; }
             [J("texture_new")] public string TextureNew { get; set; }
             [J("particles_new")] public string ParticlesNew { get; set; }
             [J("materials")] public Materials Materials { get; set; }
@@ -107,8 +107,8 @@ namespace SeaBotCore.Data.Defenitions
 
         public class Material
         {
-            [J("id")] public long Id { get; set; }
-            [J("amount")] public long Amount { get; set; }
+            [J("id")] public int Id { get; set; }
+            [J("amount")] public int Amount { get; set; }
         }
 
         public class ParticleNew
@@ -118,7 +118,7 @@ namespace SeaBotCore.Data.Defenitions
 
         public class Particle
         {
-            [J("particle_id")] public long ParticleId { get; set; }
+            [J("particle_id")] public int ParticleId { get; set; }
             [J("data")] public string Data { get; set; }
         }
 
@@ -129,10 +129,10 @@ namespace SeaBotCore.Data.Defenitions
 
         public class ProdOutput
         {
-            [J("id")] public long Id { get; set; }
-            [J("material_id")] public long MaterialId { get; set; }
-            [J("amount")] public long Amount { get; set; }
-            [J("time")] public long Time { get; set; }
+            [J("id")] public int Id { get; set; }
+            [J("material_id")] public int MaterialId { get; set; }
+            [J("amount")] public int Amount { get; set; }
+            [J("time")] public int Time { get; set; }
 
             [J("inputs", NullValueHandling = N.Ignore)]
             public Inputs Inputs { get; set; }
@@ -151,8 +151,8 @@ namespace SeaBotCore.Data.Defenitions
         public class Slot
         {
             [J("type")] public string Type { get; set; }
-            [J("id")] public long Id { get; set; }
-            [J("amount")] public long Amount { get; set; }
+            [J("id")] public int Id { get; set; }
+            [J("amount")] public int Amount { get; set; }
         }
     }
 }
